Add renaming sorts of SoftWare Heritage persistent IDentifiers
See
https://docs.softwareheritage.org/devel/swh-model/persistent-identifiers.html
for the technical details of that spec.

These remaining four sorts of objects coincide with `git-raw`, which is
already in this table. However, adding them here is not redundant,
because the deserialization direction. While a git-raw CID may legally
point to any sort of git object, the relevant SWHIDs specify in the
identifier what sort of git object is expected, and if it points to a
different type of valid git object, deserialization fails. That means
there is no way to losslessly convert a git-coinciding SWHID into a
git-raw CID.